hotfix for merging and creating tables
There was an issue while checking if the table exists- it returned null dtype as int64 1 value and therefor, didnt create the table even though it didnt exists. In addition, there was missing logic for metrics table for creating a partitioned one. Lastly, There were 2 major fileds missing in the dimensions dataframe year,date for partioning